svn-status: Avoid Subversion warning when misinterpreting PATH argument as PATTERN in case of empty status in the PATH working copy
> svn: warning: W155007: '...' is not a working copy

When svn status PATH returns nothing, this can either be a case for the extended svn status PATTERN syntax, or the working copy at PATH actually has no changes.
Perform an svn info (with PATH / PATTERN omitted) to ensure that the CWD actually is a working copy, and only then attempt to get the filtered status there.