svn-changes: Fix handling of multiple -c arguments
These must be passed as a comma-separated list of (negative) revisions to
    svn merge [-c M[,N...] | -r N:M ...]

So, instead of collecting them as $changes elements, append to the single (optional, in case -r is used) $changes element.
Adapt the conversion into a list of revisions for the commit message, too. Here, we need to split on comma after removing the - prefix.